// src/app/sermaluc/hero/interfaces/heroe.interface.ts
// Generated by https://quicktype.io

export interface Hero {
    id?:               string;
    superhero:        string;
    publisher:        Publisher;
    alter_ego:        string;
    first_appearance: string;
    characters:       string;
    alt_image? : string;
}

export enum Publisher {
    DCComics = "DC Comics",
    MarvelComics = "Marvel Comics",
}